+++ examples/SX127x_examples/LowMemory/9_LoRa_LowMemory_RX/9_LoRa_LowMemory_RX.ino
@@ -130,7 +130,7 @@ void packet_is_Error()
 
 void printpacketDetails()
 {
-  int8_t  PacketRSSI;              //RSSI of received packet
+  int16_t PacketRSSI;              //RSSI of received packet
   int8_t  PacketSNR;               //signal to noise ratio of received packet
   
   PacketRSSI = LoRa.readPacketRSSI();
